python/requirements.txt
requests
cassandra-driver==3.28.0
geomet==0.2.1.post1